refactor structure + remove comments
